feat(user): Add user management component and database tables (#1)
* Add user management component and database tables

This commit includes the creation of several database tables namely products, users, payments, orders and orders_products. It also adds a new user management component with services for JWT token generation and validation, user login, registration, update and deletion functionalities. Exception handling and proper security configurations have also been incorporated in the code.

* fix: Add sonar.java.binaries path in build.yaml

The SonarQube Java binaries path is added to the GitHub workflows. This modification specifically in the sonar args section is necessary for the proper identification and analysis of the built Java files.

Signed-off-by: Robin JOSEPH <rjrdutg@gmail.com>

* fix: Add sonar.java.binaries path in build.yaml

The SonarQube Java binaries path is added to the GitHub workflows. This modification specifically in the sonar args section is necessary for the proper identification and analysis of the built Java files.

Signed-off-by: Robin JOSEPH <rjrdutg@gmail.com>

---------

Signed-off-by: Robin JOSEPH <rjrdutg@gmail.com>